Fix FecReader block duration estimation #688
The whole combination of factor that enabled
the issue are still unclear. The fix is just in
simplifying the logic by disabling estimator for
the blocks with absent first packet, which
should not be a great deal.
Apparently the issue could take place
during start-up transition, and it
basically lead to wrong block length estimation
with two times greater value than correct one.